Move feature to back end only on a loop
