+++ static/sass/layout/_header.scss
@@ -72,25 +72,6 @@
         margin-left: 0.5em;
       }
 
-			&[href="#rarity_update"] {
-				@include icon('\f021');
-				-webkit-tap-highlight-color: rgba(0,0,0,0);
-
-				&:before {
-					margin: 0 0.5em 0 0;
-				}
-			}
-
-			& + a[href="#rarity_update"]:last-child {
-				border-left: solid 1px _palette(accent2, border);
-				padding-left: 1.25em;
-				margin-left: 0.5em;
-			}
-
-			&[href="#rarity_update"].no-before:before {
-			    content: none;
-			}
-
       @include breakpoint(small) {
         padding: 0 0.5em;
