+++ static/sass/mobile.scss
@@ -33,7 +33,7 @@ li {
   top: 20px;
   padding-bottom: 10px;
 }
-#nav, #update_rarities {
+#nav {
   position: absolute;
   top: 55px;
   left: 10px;
