+++ templates/statistics.html
@@ -42,10 +42,6 @@
         <header id="header">
             <a href="#nav">Options</a>
             <h1><a href="./">Pokémon Go Map</a></h1>
-            <div id="loading_small" class="loading" style="float: right;display: none;">
-                <img src="{{ url_for('static', filename='images/loading.gif').lstrip('/') }}" alt="Loading"/>
-            </div>
-            <a href="#rarity_update" id="update_rarities" class="raritiesButton" style="float: right;"><span class="label">Update Rarities</span></a>
         </header>
         <nav id="nav">
           <div class="ui-accordion-content">
